Issue: WSW-17509 MISRA Rule 10.3 patch 11.1
[Issue]
fn wifi_mem_malloc_cmdrespbuf should use size_t as argument

[Fix]
size_t as argument used for wifi_mem_malloc_cmdrespbuf

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.

[Open Issues/Todo]
Other 10.3 Issues.

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

+++ wifidriver/wifi-internal.h
@@ -306,7 +306,7 @@ void *wifi_mem_malloc_cmdrespbuf(void);
  *
  *
  */
-void *wifi_malloc_eventbuf(int size);
+void *wifi_malloc_eventbuf(size_t size);
 void wifi_free_eventbuf(void *buffer);
 
 void wifi_uap_handle_cmd_resp(HostCmd_DS_COMMAND *resp);
